Extract Docker exceptions to separate file
